+++ app-tv/src/test/kotlin/dev/quiblo/tv/ui/search/TvSearchRestingCentreTest.kt
@@ -195,6 +195,7 @@ class TvSearchRestingCentreTest {
                     onOpen = { _, _ -> },
                     onQueryChange = {},
                     onSelectGenre = {},
+                    onSelectYear = {},
                     onToggleIncludeHidden = {},
                     onToggleIncludeLive = {},
                     onToggleAdvanced = {},